+++ src/main/java/com/hfr/inventory/container/ContainerMachineBuilder.java
@@ -17,7 +17,9 @@ public class ContainerMachineBuilder extends Container {
 
 		diFurnace = tedf;
 
-		this.addSlotToContainer(new Slot(tedf, 0, 26, 53));
+		this.addSlotToContainer(new Slot(tedf, 0, 152, 62));
+		for(int i=0;i<3;i++) for(int j=0;j<9;j++)
+			this.addSlotToContainer(new Slot(tedf, 1+j+i*9, 8+j*18, 8+i*18));
 		
 		for(int i = 0; i < 3; i++)
 		{
@@ -47,12 +49,12 @@ public class ContainerMachineBuilder extends Container {
 			ItemStack var5 = var4.getStack();
 			var3 = var5.copy();
 
-			if (par2 <= 1) {
-				if (!this.mergeItemStack(var5, 2, this.inventorySlots.size(), true)) {
+			if (par2 < 28) {
+				if (!this.mergeItemStack(var5, 28, this.inventorySlots.size(), true)) {
 					return null;
 				}
 			} else {
-				if (!this.mergeItemStack(var5, 0, 1, false))
+				if (!this.mergeItemStack(var5, 1, 28, false))
 					return null;
 			}